refactor: pr 리뷰 반영, n+1 조회되는 문제를 fetch join 쿼리로 변경
