+++ src/main/java/doritos/doriroom/atlas/exception/InsufficientAtlasLevelException.java
@@ -1,4 +1,4 @@
-package doritos.doriroom.atlas.excetion;
+package doritos.doriroom.atlas.exception;
 
 import doritos.doriroom.global.exception.ApiException;
 import org.springframework.http.HttpStatus;
