+++ src/main/java/doritos/doriroom/atlas/repository/UserAtlasRepository.java
@@ -7,6 +7,7 @@ import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,5 +20,7 @@ public interface UserAtlasRepository extends JpaRepository<UserAtlas, UUID> {
     @Query("SELECT ua FROM UserAtlas ua WHERE ua.user = :user AND ua.atlas = :atlas")
     Optional<UserAtlas> findByUserAndAtlasWithLock(User user, Atlas atlas); // 락으로 동시 발생 제어(경험치 추가 시)
 
-    List<UserAtlas> findByUser(User user); // 유저의 모든 지역도감 조회 시
+//    List<UserAtlas> findByUser(User user); // 유저의 모든 지역도감 조회 시
+    @Query("SELECT ua FROM UserAtlas ua JOIN FETCH ua.atlas WHERE ua.user = :user") // fetch join 하여 UserAtlas와 Atlas를 전부 조회
+    List<UserAtlas> findByUserWithAtlas(@Param("user") User user);
 }
