+++ src/main/java/doritos/doriroom/atlas/repository/UserAtlasRewardRepository.java
@@ -11,7 +11,9 @@ import java.util.UUID;
 
 public interface UserAtlasRewardRepository extends JpaRepository<UserAtlasReward, UUID> {
 
-    List<UserAtlasReward> findByUser(User user); // 특정 유저가 받은 모든 보상 아이템 조회
+//    List<UserAtlasReward> findByUser(User user); // 특정 유저가 받은 모든 보상 아이템 조회
+    @Query("SELECT uar FROM UserAtlasReward uar JOIN FETCH uar.atlasReward ar JOIN FETCH ar.atlas WHERE uar.user = :user")
+    List<UserAtlasReward> findByUserWithDetails(@Param("user") User user); // fetch join 적용 AtlasReward와 Atlas을 함께 조회
 
     // 특정 유저가 특정 지역에서 받은 보상 아이템들 조회
     @Query("SELECT uar FROM UserAtlasReward uar " +
