+++ src/main/java/doritos/doriroom/challenge/exception/ChallengeStatusException.java
@@ -1,4 +1,4 @@
-package doritos.doriroom.challenge.excetion;
+package doritos.doriroom.challenge.exception;
 
 import doritos.doriroom.global.exception.ApiException;
 import org.springframework.http.HttpStatus;
