+++ src/main/java/doritos/doriroom/quiz/service/QuizService.java
@@ -4,8 +4,8 @@ import doritos.doriroom.challenge.domain.challenge.Challenge;
 import doritos.doriroom.challenge.domain.userchallenge.ChallengeStatus;
 import doritos.doriroom.challenge.domain.userchallenge.UserChallenge;
 import doritos.doriroom.challenge.dto.ChallengeRewardDto;
-import doritos.doriroom.challenge.excetion.ChallengeNotFoundException;
-import doritos.doriroom.challenge.excetion.ChallengeStatusException;
+import doritos.doriroom.challenge.exception.ChallengeNotFoundException;
+import doritos.doriroom.challenge.exception.ChallengeStatusException;
 import doritos.doriroom.challenge.repository.ChallengeRepository;
 import doritos.doriroom.challenge.repository.UserChallengeRepository;
 import doritos.doriroom.quiz.domain.Question;
